Pulled out `SummaryCollector` for all the logic to construct summary when the data arrived.
